// packages/components/Trigger/interface.ts
export interface TriggerProps {
// ...
  position?:
    | 'top'
    | 'tl'
    | 'tr'
    | 'bottom'
    | 'bl'
    | 'br'
    | 'left'
    | 'lt'
    | 'lb'
    | 'right'
    | 'rt'
    | 'rb';
// ...
  boundaryDistance?:
    | { left?: number; bottom?: number }
    | { left?: number; top?: number }
    | { right?: number; bottom?: number }
    | { right?: number; top?: number };
// ...
}